+++ kernel/pkg/README.md
@@ -1,0 +1,13 @@
+# 操作メモ
+
+コンパイル
+
+```
+cargo run
+```
+
+WASM のビルド
+
+```
+dc exec rust wasm-pack build
+```
