+++ kernel/pkg/rawp_kernel.d.ts
@@ -1,0 +1,10 @@
+/* tslint:disable */
+/* eslint-disable */
+/**
+* @param {number} n_of_people
+* @param {number} n_of_group
+* @param {number} n_of_times
+* @param {bigint} seed
+* @returns {string}
+*/
+export function resolve_by_sa(n_of_people: number, n_of_group: number, n_of_times: number, seed: bigint): string;
