+++ kernel/pkg/rawp_kernel.js
@@ -1,0 +1,2 @@
+import * as wasm from "./rawp_kernel_bg.wasm";
+export * from "./rawp_kernel_bg.js";
